osleridgenservice/requirements.txt
Django
djangorestframework
drf-writable-nested
pykafka
